Inline action definition with implementation
If you look in the diff, you'll see a deleted comment about the relative
simplicity of keeping types and implementations separate for an action
signature of this complexity. Two things have changed since I wrote that
comment:
- The error type signatures became far more manageable.
- I learned how to not be a TypeScript n00b.

The second point contributed the most. I've since learned how to write
inline overloads and I've improved my type narrowing skills.

-----

The reason for the change: As you would probably guess, keeping types
and implementations separate allowed them to continually drift. I keep
noticing instances where the type is correct and the implementation is
wrong, and vice versa. I'm ready to give inline types another shot.
