+++ src/actions/create-action.ts
@@ -1,9 +1,49 @@
-import { ActionConstant } from '../types/actions';
-import { CreateAction } from '../types/create-action';
 import createAsyncAction from './create-async-action';
 import bindActionType from './bind-action-type';
 import validateActionType from './validate-action-type';
 import { isKnownError } from '../utils/errors';
+import {
+  ActionConstant,
+  CoercibleAction,
+  ActionSuccess,
+  ActionFailure,
+  VoidAction,
+} from '../types/actions';
+
+/**
+ * Returns a function which generates actions of the given type.
+ * @param type A unique name which describes the action.
+ * @param effect Something to run when the action is invoked. Whatever it
+ * returns becomes the action payload.
+ * @example
+ * createAction('settings/load-theme', () => {
+ *   return localStorage.getItem('theme')
+ * })
+ */
+export default function createAction(
+  type: ActionConstant,
+): CoercibleAction<[], ActionIterator<VoidAction>>;
+
+// No effect. Just pass a payload.
+export default function createAction<Payload>(
+  type: ActionConstant,
+): CoercibleAction<[Payload], ActionIterator<ActionSuccess<Payload>>>;
+
+// No arguments.
+export default function createAction<Effect extends () => any>(
+  type: ActionConstant,
+  effect: Effect,
+): CoercibleAction<[], ActionOutcomesForEffect<Effect>>;
+
+// At least one argument.
+export default function createAction<
+  Effect extends (arg: any, ...args: any) => any
+>(
+  type: ActionConstant,
+  effect: Effect,
+): Effect extends (arg: infer T, ...args: any) => any
+  ? CoercibleAction<[T], ActionOutcomesForEffect<Effect>>
+  : never;
 
 // Synchronous action creator.
 //
@@ -11,10 +51,7 @@ import { isKnownError } from '../utils/errors';
 // createAction('add-file', file => {
 //   return { url: URL.createObjectURL(file) }
 // })
-const createAction = (
-  actionType: ActionConstant,
-  effect?: (...args: any) => any,
-) => {
+export default function createAction(actionType: any, effect?: any) {
   validateActionType(actionType);
 
   function executeEffectAndReturnAction(input: any) {
@@ -47,7 +84,7 @@ const createAction = (
       }
     }
   });
-};
+}
 
 // Asynchronous action creator.
 //
@@ -57,7 +94,9 @@ const createAction = (
 // })
 createAction.async = createAsyncAction;
 
-// These types become completely unmaintainable when you try to write them
-// with the implementation. It's *significantly* easier to keep them in
-// a separate file and re-cast the implementation to a more precise type.
-export default (createAction as unknown) as CreateAction;
+type ActionIterator<Action> = Generator<Action, Action>;
+
+// If an effect is provided, we have to assume it can fail.
+type ActionOutcomesForEffect<
+  Effect extends (...args: any[]) => any
+> = ActionIterator<ActionSuccess<ReturnType<Effect>> | ActionFailure<unknown>>;
